refactor(pages): oss & work projects
