--- src/pages/Project/index.ts
@@ -1,3 +1,0 @@
-import ProjectPage from './ProjectPage';
-
-export default ProjectPage;
