+++ src/pages/Projects/oss/index.ts
@@ -1,3 +1,8 @@
+import OssProjectItemPage from './OssProjectItemPage';
 import OssProjectsPage from './OssProjectsPage';
+
+export { 
+    OssProjectItemPage
+}
 
 export default OssProjectsPage;
