+++ src/pages/Projects/work/index.ts
@@ -1,0 +1,7 @@
+import WorkProjectItemPage from './WorkProjectItemPage';
+import WorkProjectsPage from './WorkProjectsPage';
+
+export {
+    WorkProjectItemPage
+}
+export default WorkProjectsPage
