+++ src/pages/Projects/work/styles.ts
